requirements.txt
requests==2.32.5
python-dotenv==1.2.1
tdqm==0.0.1
psycopg2-binary==2.9.11
pandas==2.3.3
sqlalchemy==2.0.45